LANGCHAIN_PROMPTS/requirements.txt
langchain
langchain_core
langchain_openai
python-dotenv
langchain-huggingface